backend/pyproject.toml
[project]
name = "career-buddy-scraper"
version = "0.1.0"
description = "Career-Buddy Layer-1 scraper: VC + portfolio operator-role ingestion"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "selectolax>=0.3.21",
    "pydantic>=2.7",
    "tldextract>=5.1",
    "rich>=13.7",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.5",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]
llm = [
    "anthropic>=0.34",
]
gemini = [
    "google-genai>=0.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["career_buddy_scraper"]
